+++ cli/cmd/base_tools.go
@@ -2,12 +2,36 @@ package cmd
 
 import (
 	"fmt"
+	"github.com/pkg/errors"
 	"io"
+	"io/ioutil"
 	"os"
 
 	"github.com/gogo/protobuf/jsonpb"
 	"github.com/gogo/protobuf/proto"
 )
+
+func readJsonFileToProtobuf(filename string, message proto.Message) error {
+	json, err := readFileToString(filename)
+	if err != nil {
+		return errors.Wrap(err, "error reading " + filename)
+	}
+
+	if err := jsonpb.UnmarshalString(json, message); err != nil {
+		return errors.Wrap(err, "error parsing JSON file " + filename)
+	}
+
+	return nil
+}
+
+func readFileToString(filename string) (string, error) {
+	bytes, err := ioutil.ReadFile(filename)
+	if err != nil {
+		return "", err
+	}
+
+	return string(bytes), nil
+}
 
 func printProtoMessageAsJSON(out io.Writer, pb proto.Message) error {
 	m := jsonpb.Marshaler{
